Including Pong Button and opponent
I was able to make the bottom two buttons work. I did this by declaring two void functions that makes either p2 true and opponent false, or vice versa. Then I made an if statement checking if the booleans were true and activated the P2 or the opponent. I will next include it so that way it is being instantiated instead of being set active