Pong Prototype 1 is a pass!
I was able to build the game for IOS and upload it to Xcode successfully! I spent the next half of my time making an exit button for the player and opponent, as well as figuring out why the score text is not appearing for the game. We'll just have to figure it out next time...