Fixed Split Pong/added more to it/ gave impossible pong an icon photo
I was able to fix the issue with the split game mode by instead of disabling and enabling the entire gameobject of the opponent, I grabbed its renderer and flicked that on and off. I am still not sure why I am not able to disable the object but I have to move forward. I then made it so every time either the opp or the player scores, they go back a step in the state functionality of paddle splitting. I then drew an evil monster face by referencing an image on the internet using sprites in aa new scene. It did not turn out exactly how I wanted it, but I think it is pretty funny and cool... More to come on the next development phaase!