I was able to fix The Multiple Gamemode
I went into the scripting and was able to take ball_count (set to 0) and increase it by how many I wanted balls to be spawned into the game. This variable checked how many balls were being instantiated. I also put all the if statement functionality originally happening on the collision method and put it into one function. That way it can check if the limit is over or not, and depending on what is returned true, the function will execute. I'm not sure if I fixed the issue because I don't want it to still continue to spawn but we'll see.... more on the next dev day!