Some Late night Programming
I played a bit of my beta that I built last night and noticed some bugs. I had to make some adjustments to some of the game settings on the paddles and the balls just to make sure they were working well. I then went to profiler and looked at the garbage collection to make sure nothing was eating too much CPU. When I played the built game I noticed that it was a little laggy and noticed it has an FPD of roughly 15-20, which is not amazing, especially for a pong game. I'll have to watch some videos to see how I can fix this. I then had an idea for a pretty low-end exit button. I copied the game scene, and essentially the exit button is jumping between two of the same scenes, so just duplicates, without the player knowing. I am using the change scene script, and bouncing back between scenes 0-1, and i tested it tonight and it works great. this isn't ideal for a game usually, but for a game that only has one game scene with many sub game scenes layered, it works great...