Background sound/size game fixes
I made sure I could navigate throughout the entire game and get the different songs to change and work smoothly. I had to make sure some of the music stopped and others didn't. It gave me an awesome idea to produce some more background music depending on which game mode you decide to pick. I then worked on the size game mode and realized that when the ball is the smallest, it needs to go to the game over scene... I tried doing that by adding an if statement but noticed that the paddle was not getting to a value of 0, but something just a tad bit bigger (2.980232e-08f). So it still does not work but I need to test it....