Player selection input/bug fixes
I created a quick program that I attached to the MainMenu script that checks if the user clicks on the opp or p2 buttons. Depending on which they clicked, I have two images of pong paddles that become less transparent if one is selected, that way you can tell which button you selected. I also went in and made sure it was consistent between both scenes. One final thing I did was grow the exit button just so it is a little more visible for the player...